fix(sdk): throw error if invalid deployments
