+++ components/clarinet-cli/Cargo.toml
@@ -28,9 +28,6 @@ log = { version = "=0.4.17", features = ["serde"] }
 signal-hook-registry = "1.4.0"
 secure_tempfile = { version = "3.8.0", package = "tempfile" }
 tokio-util = { version = "0.7.1", features = ["io"], optional = true }
-clarity_repl = { package = "clarity-repl", path = "../clarity-repl", features = [
-    "cli",
-] }
 libsecp256k1 = "0.7.0"
 hmac = "0.12.0"
 pbkdf2 = { version = "0.11.0", features = ["simple"], default-features = false }
@@ -60,18 +57,22 @@ mac_address = { version = "1.1.2", optional = true }
 tower-lsp = { version = "0.19.0", optional = true }
 hex = "0.4.3"
 serde_yaml = "0.8.23"
-clarinet-files = { path = "../clarinet-files", features = ["cli"] }
-clarity-lsp = { path = "../clarity-lsp", features = ["cli"] }
-clarinet-deployments = { path = "../clarinet-deployments", features = ["cli"] }
-hiro-system-kit = { path = "../hiro-system-kit" }
-clarinet-utils = { path = "../clarinet-utils" }
-stacks-network = { path = "../stacks-network" }
 num_cpus = "1.13.1"
 mio = "0.8"
 similar = "2.1.0"
 crossbeam-channel = "0.5.6"
 chrono = "0.4.31"
 sha2 = "0.10.0"
+
+clarity_repl = { package = "clarity-repl", path = "../clarity-repl", features = [
+    "cli",
+] }
+clarinet-files = { path = "../clarinet-files", features = ["cli"] }
+clarity-lsp = { path = "../clarity-lsp", features = ["cli"] }
+clarinet-deployments = { path = "../clarinet-deployments", features = ["cli"] }
+hiro-system-kit = { path = "../hiro-system-kit" }
+clarinet-utils = { path = "../clarinet-utils" }
+stacks-network = { path = "../stacks-network" }
 
 [dependencies.tui]
 version = "0.18.0"
