+++ components/clarinet-deployments/Cargo.toml
@@ -4,6 +4,7 @@ version = "1.0.1"
 edition = "2021"
 
 [dependencies]
+colored = "2.0.4"
 serde = "1"
 serde_json = "1"
 serde_derive = "1"
