+++ components/clarinet-sdk/tests/fixtures/contracts/invalid.clar
@@ -1,0 +1,5 @@
+(define-data-var count uint u0)
+
+(define-read-only  (get-count)
+  (ok { count: (var-get count) })
+)) ;; extra `)`
