+++ components/clarinet-sdk/vitest.config.js
@@ -3,6 +3,7 @@ import { defineConfig } from "vite";
 
 export default defineConfig({
   test: {
+    singleThread: true,
     include: ["./tests/**/*.test.ts", "./vitest-helpers/tests/**/*.test.ts"],
   },
 });
